feat: added deeplink provider that is used by phantom and solflare
