+++ packages/solana/src/index.ts
@@ -2,7 +2,8 @@
 export { SolanaAdapter } from './adapter';
 
 // Types
-export type { PhantomConnectorConfig } from './types';
+export type { PhantomConnectorConfig, SolflareConnectorConfig } from './types';
 
 // Connectors
 export { PhantomConnector } from './connectors/PhantomConnector';
+export { SolflareConnector } from './connectors/SolflareConnector';
